# data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI营销战略执行智能体",
    "description": "用AI帮小企业把营销战略翻译成每天的执行动作",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/plan/weekly": {
      "get": {
        "tags": [
          "周计划"
        ],
        "summary": "Get Weekly Plan V3",
        "description": "[V1.0] 获取周执行计划（文档6.3.4节 SevenDayPlan 结构）。",
        "operationId": "get_weekly_plan_v3_api_plan_weekly_get",
        "parameters": [
          {
            "name": "week_number",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "周数 1-12（对应季度12周）",
              "default": 1,
              "title": "Week Number"
            },
            "description": "周数 1-12（对应季度12周）"
          },
          {
            "name": "business_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Business Id"
            }
          },
          {
            "name": "diagnosis_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Diagnosis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "响应状态码，0表示成功",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息",
            "default": "ok"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data",
            "description": "响应数据载荷"
          }
        },
        "type": "object",
        "title": "PlanResponse",
        "description": "周计划统一响应格式。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}